Move feeding details to Menu component
